+++ packages/edit/i18n/it.js
@@ -11,5 +11,7 @@ i18n.map('it', {
 	btn_delplace: "Elimina questo luogo",
 
 	label_editplaces: "Modifiche",
+	label_owner: "Creatore",
+
 	edit_error_editing: "Modifica ogni campo e chiudi per tornare alla vista pubblica",	
 });